requirements.txt
django==3.2.10
argon2-cffi==21.2.0
python-decouple==3.5
dj-database-url==0.5.0
django-behaviors==0.5.1
django-simple-history==3.0.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
django-oauth-toolkit==1.5.0
django-filter==21.1
markdown==3.3.6
psycopg2==2.6.1
gunicorn==19.4.1

